Move runGovernance command to Governance module
