+++ cardano-cli/src/Cardano/CLI/EraBased/Run/Governance.hs
@@ -3,11 +3,14 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Cardano.CLI.EraBased.Run.Governance
-  ( runGovernanceMIRCertificatePayStakeAddrs
+  ( runGovernanceCmds
+
+  , runGovernanceMIRCertificatePayStakeAddrs
   , runGovernanceMIRCertificateTransfer
   ) where
 
@@ -15,14 +18,58 @@ import           Cardano.Api
 import qualified Cardano.Api.Ledger as Ledger
 import           Cardano.Api.Shelley
 
+import qualified Cardano.CLI.EraBased.Commands.Governance as Cmd
+import           Cardano.CLI.EraBased.Run.Governance.Actions
+import           Cardano.CLI.EraBased.Run.Governance.Committee
+import           Cardano.CLI.EraBased.Run.Governance.DRep
+import           Cardano.CLI.EraBased.Run.Governance.GenesisKeyDelegationCertificate
+import           Cardano.CLI.EraBased.Run.Governance.Poll
+import           Cardano.CLI.EraBased.Run.Governance.Vote
 import           Cardano.CLI.Types.Common
+import           Cardano.CLI.Types.Errors.CmdError
 import           Cardano.CLI.Types.Errors.GovernanceCmdError
 import qualified Cardano.Ledger.Shelley.TxBody as Shelley
 
 import           Control.Monad
 import           Control.Monad.Trans.Except (ExceptT)
 import           Control.Monad.Trans.Except.Extra
+import           Data.Function
 import qualified Data.Map.Strict as Map
+
+runGovernanceCmds :: ()
+  => Cmd.GovernanceCmds era
+  -> ExceptT CmdError IO ()
+runGovernanceCmds = \case
+  Cmd.GovernanceMIRPayStakeAddressesCertificate w mirpot vKeys rewards out ->
+    runGovernanceMIRCertificatePayStakeAddrs w mirpot vKeys rewards out
+      & firstExceptT CmdGovernanceCmdError
+
+  Cmd.GovernanceMIRTransfer w ll oFp direction ->
+    runGovernanceMIRCertificateTransfer w ll oFp direction
+      & firstExceptT CmdGovernanceCmdError
+
+  Cmd.GovernanceGenesisKeyDelegationCertificate sta genVk genDelegVk vrfVk out ->
+    let stb = shelleyToAlonzoEraToShelleyToBabbageEra sta in
+    runGovernanceGenesisKeyDelegationCertificate stb genVk genDelegVk vrfVk out
+      & firstExceptT CmdGovernanceCmdError
+
+  Cmd.GovernanceCommitteeCmds cmds ->
+    runGovernanceCommitteeCmds cmds
+      & firstExceptT CmdGovernanceCommitteeError
+
+  Cmd.GovernanceActionCmds cmds ->
+    runGovernanceActionCmds cmds
+      & firstExceptT CmdGovernanceActionError
+
+  Cmd.GovernanceDRepCmds cmds ->
+    runGovernanceDRepCmds cmds
+
+  Cmd.GovernancePollCmds cmds ->
+    runGovernancePollCmds cmds
+      & firstExceptT CmdGovernanceCmdError
+
+  Cmd.GovernanceVoteCmds cmds ->
+    runGovernanceVoteCmds cmds
 
 runGovernanceMIRCertificatePayStakeAddrs
   :: ShelleyToBabbageEra era
